+++ play/src/main/java/com/example/play/IPaly.kt
@@ -11,5 +11,5 @@ interface IPaly {
     fun resume()
     fun pause()
     fun release()
-    fun seekTo(seek: Double)
+    fun seekTo(seek: Long)
 }
